chore(group-behavior): move category creation to util
